+++ app/entity/dto/MovieTypeDTO.go
@@ -1,7 +1,7 @@
 package dto
 
 type MovieTypeDTO struct {
-	MovieTypeId uint   `json:"movie_type_id,omitempty"`
+	MovieTypeId int64  `json:"movie_type_id,omitempty"`
 	Name        string `json:"name,omitempty"`
 	Slug        string `json:"slug,omitempty"`
 }
