+++ app/entity/model/EpisodeDetail.go
@@ -1,10 +1,10 @@
 package model
 
 type EpisodeDetail struct {
-	EpisodeDetailId uint   `gorm:"primaryKey" json:"episode_detail_id"`
+	EpisodeDetailId int64  `gorm:"primaryKey" json:"episode_detail_id"`
 	Name            string `gorm:"not null" json:"name"`
-	EpisodeId       uint   `gorm:"index:,not null" json:"episode_id"`
-	EpisodeTypeId   uint   `gorm:"index:,not null" json:"episode_type_id"`
+	EpisodeId       int64  `gorm:"index:,not null" json:"episode_id"`
+	EpisodeTypeId   int64  `gorm:"index:,not null" json:"episode_type_id"`
 	Status          int    `gorm:"index:,default:1" json:"status"`
 	BaseModel
 }
