+++ app/entity/model/MovieTag.go
@@ -1,6 +1,6 @@
 package model
 
 type MovieTag struct {
-	MovieId uint `gorm:"index:,not null" json:"movie_id"`
-	TagId   uint `gorm:"index:,not null" json:"tag_id"`
+	MovieId int64 `gorm:"index:,not null" json:"movie_id"`
+	TagId   int64 `gorm:"index:,not null" json:"tag_id"`
 }
