+++ app/entity/model/MovieType.go
@@ -1,7 +1,7 @@
 package model
 
 type MovieType struct {
-	MovieTypeId uint   `gorm:"primaryKey" json:"movie_type_id"`
+	MovieTypeId int64  `gorm:"primaryKey" json:"movie_type_id"`
 	Name        string `gorm:"not null" json:"name"`
 	Slug        string `gorm:"not null" json:"slug"`
 	Status      int    `gorm:"index:,default:1" json:"status"`
