+++ pkg/validator/MovieGenreValidator.go
@@ -4,7 +4,7 @@ import (
 	"github.com/xdorro/golang-fiber-base-project/app/entity/model"
 )
 
-func ExistGenreIdInGenres(genreId uint, genres []model.Genre) bool {
+func ExistGenreIdInGenres(genreId int64, genres []model.Genre) bool {
 	for _, genre := range genres {
 		if genre.GenreId == genreId {
 			return true
@@ -14,7 +14,7 @@ func ExistGenreIdInGenres(genreId uint, genres []model.Genre) bool {
 	return false
 }
 
-func ExistGenreIdInGenreIds(genreId uint, genreIds []uint) bool {
+func ExistGenreIdInGenreIds(genreId int64, genreIds []int64) bool {
 	for _, genre := range genreIds {
 		if genre == genreId {
 			return true
